dev: split main module move all functions into individual modules
