+++ download.go
@@ -1,0 +1,31 @@
+package main
+
+import (
+	"io"
+	"net/http"
+	"os"
+)
+
+// DownloadFile will download from a given url to a file. It will
+// write as it downloads (useful for large files).
+func DownloadFile(filepath string, url string) error {
+
+	// Get the data
+	resp, err := http.Get(url)
+	if err != nil {
+		return err
+	}
+	defer resp.Body.Close()
+
+	// Create the file
+	out, err := os.Create(filepath)
+	if err != nil {
+		return err
+	}
+	defer out.Close()
+
+	// Write the body to file
+	_, err = io.Copy(out, resp.Body)
+	return err
+
+}
