+++ hash.go
@@ -1,0 +1,30 @@
+package main
+
+import (
+	"hash/crc32"
+	"io"
+	"os"
+)
+
+func getHash(filename string) (uint32, error) {
+	// open the file
+	f, err := os.Open(filename)
+	if err != nil {
+		return 0, err
+	}
+	// remember to always close opened files
+	defer f.Close()
+
+	// create a hasher
+
+	h := crc32.NewIEEE()
+	// copy the file into the hasher
+	// - copy takes (dst, src) and returns (bytesWritten, error)
+	_, err = io.Copy(h, f)
+	// we don't care about how many bytes were written, but we do want to
+	// handle the error
+	if err != nil {
+		return 0, err
+	}
+	return h.Sum32(), nil
+}
